refactor: consolidate test setup fixtures in packages/server
Drop 5 unused fixture factories from `test-fixtures.ts` (createTestWatcher,
createTestWatcherTemplate, createTestWatcherWindow, createTestActionRun,
createTestClassifier — zero importers) and merge `test-workspace.ts` into
`test-mcp-client.ts` since the workspace builder is just a thin wrapper
over TestApiClient. Net -323 lines, one fewer file in `__tests__/setup/`.
