+++ packages/server/src/__tests__/setup/test-fixtures.ts
@@ -8,7 +8,6 @@
 import { serializeSigned } from 'hono/utils/cookie';
 import { generateSecureToken, hashToken } from '../../auth/oauth/utils';
 import { pgTextArray } from '../../db/client';
-import { nextRunAt } from '../../utils/cron';
 import { generateSlug } from '../../utils/entity-management';
 import { getTestDb } from './test-db';
 
@@ -655,324 +654,6 @@ export async function createTestEvent(options: {
   return {
     id: Number(inserted.id),
     origin_id: String(inserted.origin_id),
-  };
-}
-
-// ============================================
-// Watcher + Version Fixtures
-// ============================================
-
-interface TestWatcherTemplate {
-  id: string;
-  slug: string;
-}
-
-export async function createTestWatcherTemplate(options: {
-  slug: string;
-  name: string;
-  version?: number;
-  prompt?: string;
-  output_schema?: Record<string, any>;
-  organization_id?: string | null;
-  entity_id?: number;
-}): Promise<TestWatcherTemplate> {
-  const sql = getTestDb();
-
-  // Ensure seed user exists for FK constraint
-  await sql`
-    INSERT INTO "user" (id, name, email, username, "emailVerified", "createdAt", "updatedAt")
-    VALUES ('test-seed-user', 'Test Seed User', 'test-seed-user@example.com', 'test-seed-user', true, NOW(), NOW())
-    ON CONFLICT (id) DO NOTHING
-  `;
-
-  // Create the watcher row
-  const [{ next_id }] = await sql`
-    SELECT COALESCE(MAX(id), 0) + 1 AS next_id FROM watchers
-  `;
-
-  const entityIds = options.entity_id ? `{${options.entity_id}}` : null;
-
-  const [inserted] = await sql`
-    INSERT INTO watchers (
-      id, slug, name, description,
-      organization_id, created_by,
-      entity_ids, version, status,
-      created_at, updated_at
-    ) VALUES (
-      ${next_id},
-      ${options.slug},
-      ${options.name},
-      ${'Test watcher'},
-      ${options.organization_id ?? null},
-      'test-seed-user',
-      ${entityIds ? sql`${entityIds}::bigint[]` : null},
-      ${options.version ?? 1},
-      'active',
-      NOW(), NOW()
-    )
-    RETURNING id, slug
-  `;
-
-  const watcherId = Number(inserted.id);
-
-  // Create a version for this watcher
-  const [{ next_vid }] = await sql`
-    SELECT COALESCE(MAX(id), 0) + 1 AS next_vid FROM watcher_versions
-  `;
-
-  const [versionRow] = await sql`
-    INSERT INTO watcher_versions (
-      id, watcher_id, version, name, description, prompt,
-      extraction_schema, change_notes,
-      created_by, created_at
-    ) VALUES (
-      ${next_vid},
-      ${watcherId},
-      ${options.version ?? 1},
-      ${options.name},
-      ${'Test watcher'},
-      ${options.prompt ?? 'Analyze the content for {{entities}}'},
-      ${sql.json(options.output_schema ?? { type: 'object', properties: { summary: { type: 'string' } } })},
-      'Initial version',
-      'test-seed-user',
-      NOW()
-    )
-    RETURNING id
-  `;
-
-  // Set current version ID
-  await sql`
-    UPDATE watchers
-    SET current_version_id = ${versionRow.id}
-    WHERE id = ${watcherId}
-  `;
-
-  return {
-    id: String(watcherId),
-    slug: options.slug,
-  };
-}
-
-// ============================================
-// Watcher Fixtures (attach entity to existing watcher)
-// ============================================
-
-interface TestWatcher {
-  id: number;
-  entity_id: number;
-  version_id: number;
-}
-
-export async function createTestWatcher(options: {
-  entity_id: number;
-  template_id: string;
-  organization_id: string;
-  status?: string;
-  schedule?: string;
-  agent_id?: string;
-  scheduler_client_id?: string;
-  filters?: Record<string, any>;
-}): Promise<TestWatcher> {
-  const sql = getTestDb();
-
-  const watcherId = Number(options.template_id);
-
-  // Update the watcher to attach the entity and set options
-  await sql`
-    UPDATE watchers
-    SET entity_ids = ${`{${options.entity_id}}`}::bigint[],
-        organization_id = ${options.organization_id},
-        status = ${options.status ?? 'active'},
-        schedule = ${options.schedule ?? null},
-        next_run_at = ${options.schedule ? nextRunAt(options.schedule) : null},
-        agent_id = ${options.agent_id ?? null},
-        scheduler_client_id = ${options.scheduler_client_id ?? null},
-        updated_at = NOW()
-    WHERE id = ${watcherId}
-  `;
-
-  // Get current version id
-  const [w] = await sql`
-    SELECT current_version_id FROM watchers WHERE id = ${watcherId} LIMIT 1
-  `;
-
-  return {
-    id: watcherId,
-    entity_id: options.entity_id,
-    version_id: Number(w.current_version_id),
-  };
-}
-
-// ============================================
-// Watcher Window Fixtures
-// ============================================
-
-interface TestWatcherWindow {
-  id: number;
-  watcher_id: number;
-}
-
-export async function createTestWatcherWindow(options: {
-  watcher_id: number;
-  window_start?: Date;
-  window_end?: Date;
-  granularity?: string;
-  extracted_data?: Record<string, any>;
-  content_analyzed?: number;
-  model_used?: string;
-  client_id?: string;
-  run_metadata?: Record<string, any>;
-}): Promise<TestWatcherWindow> {
-  const sql = getTestDb();
-  const windowStart = options.window_start ?? new Date(Date.now() - 7 * 24 * 60 * 60 * 1000);
-  const windowEnd = options.window_end ?? new Date();
-
-  const [inserted] = await sql`
-    INSERT INTO watcher_windows (
-      watcher_id, granularity, window_start, window_end,
-      extracted_data, content_analyzed, model_used, client_id, run_metadata, created_at
-    ) VALUES (
-      ${options.watcher_id},
-      ${options.granularity ?? 'weekly'},
-      ${windowStart},
-      ${windowEnd},
-      ${sql.json(options.extracted_data ?? { summary: 'Test summary' })},
-      ${options.content_analyzed ?? 10},
-      ${options.model_used ?? 'test-model'},
-      ${options.client_id ?? null},
-      ${sql.json(options.run_metadata ?? {})},
-      NOW()
-    )
-    RETURNING id, watcher_id
-  `;
-
-  return {
-    id: Number(inserted.id),
-    watcher_id: Number(inserted.watcher_id),
-  };
-}
-
-// ============================================
-// Action Run Fixtures
-// ============================================
-
-interface TestActionRun {
-  id: number;
-  connection_id: number;
-  status: string;
-}
-
-export async function createTestActionRun(options: {
-  connection_id: number;
-  organization_id: string;
-  action_key: string;
-  status?: string;
-  approval_status?: string;
-  input?: Record<string, any>;
-  output?: Record<string, any>;
-}): Promise<TestActionRun> {
-  const sql = getTestDb();
-
-  const [inserted] = await sql`
-    INSERT INTO runs (
-      connection_id, organization_id, run_type,
-      status, approval_status,
-      action_key, action_input, action_output,
-      created_at
-    ) VALUES (
-      ${options.connection_id},
-      ${options.organization_id},
-      'action',
-      ${options.status ?? 'pending'},
-      ${options.approval_status ?? 'pending'},
-      ${options.action_key},
-      ${sql.json(options.input ?? {})},
-      ${sql.json(options.output ?? {})},
-      NOW()
-    )
-    RETURNING id, connection_id, status
-  `;
-
-  return {
-    id: Number(inserted.id),
-    connection_id: Number(inserted.connection_id),
-    status: String(inserted.status),
-  };
-}
-
-// ============================================
-// Classifier Fixtures
-// ============================================
-
-interface TestClassifier {
-  id: number;
-  slug: string;
-  current_version_id: number;
-}
-
-export async function createTestClassifier(options: {
-  organization_id: string;
-  slug: string;
-  name?: string;
-  attribute_key?: string;
-  attribute_values?: Record<string, any>;
-  entity_id?: number;
-  status?: string;
-}): Promise<TestClassifier> {
-  const sql = getTestDb();
-  const attributeKey = options.attribute_key ?? options.slug;
-
-  // Ensure seed user for created_by FK
-  await sql`
-    INSERT INTO "user" (id, name, email, username, "emailVerified", "createdAt", "updatedAt")
-    VALUES ('test-seed-user', 'Test Seed User', 'test-seed-user@example.com', 'test-seed-user', true, NOW(), NOW())
-    ON CONFLICT (id) DO NOTHING
-  `;
-
-  const [inserted] = await sql`
-    INSERT INTO event_classifiers (
-      organization_id, slug, name, attribute_key,
-      entity_id, status, created_by,
-      created_at, updated_at
-    ) VALUES (
-      ${options.organization_id},
-      ${options.slug},
-      ${options.name ?? options.slug},
-      ${attributeKey},
-      ${options.entity_id ?? null},
-      ${options.status ?? 'active'},
-      'test-seed-user',
-      NOW(), NOW()
-    )
-    RETURNING id, slug
-  `;
-
-  const classifierId = Number(inserted.id);
-
-  const attrValues = options.attribute_values ?? {
-    positive: { description: 'Positive' },
-    negative: { description: 'Negative' },
-    neutral: { description: 'Neutral' },
-  };
-
-  const [versionRow] = await sql`
-    INSERT INTO event_classifier_versions (
-      classifier_id, version, attribute_values, is_current, created_by, created_at
-    ) VALUES (
-      ${classifierId},
-      1,
-      ${sql.json(attrValues)},
-      true,
-      'test-seed-user',
-      NOW()
-    )
-    RETURNING id
-  `;
-
-  return {
-    id: classifierId,
-    slug: options.slug,
-    current_version_id: Number(versionRow.id),
   };
 }
